Renamed utils/misc.py -> utils/common.py + minor docs changes
